# go/token: improve the performance of IsKeyword