This implementation is not perfect, but there is still a little performance improvement:

The current implementation is not the most efficient. 
Obviously, there is a high probability that we need to compare twice, and we can continue to optimize the hash function.

name                          old time/op  new time/op  delta
IsKeyword-12                   261ns ± 1%   112ns ± 0%  -57.05%  (p=0.000 n=9+9)
Lookup-12                      259ns ± 2%   113ns ± 3%  -56.28%  (p=0.000 n=7+9)
IsKeyword_WithNonKeywords-12   359ns ± 3%    84ns ± 6%  -76.50%  (p=0.000 n=8+10)
Lookup_WithNonKeywords-12      353ns ± 1%    81ns ± 1%  -77.19%  (p=0.000 n=8+10)